Move control flow detection out of dace.graph
