+++ dace/codegen/targets/framecode.py
@@ -3,6 +3,7 @@ from typing import Optional, Set, Tuple
 import collections
 import dace
 import functools
+from dace.codegen import control_flow as cflow
 from dace.codegen.prettycode import CodeIOStream
 from dace.codegen.targets.common import codeblock_to_cpp
 from dace.codegen.targets.target import TargetCodeGenerator, TargetDispatcher
@@ -408,24 +409,23 @@ DACE_EXPORTED void __dace_exit_%s(%s)
 
                     for control in control_flow[edge]:
 
-                        if isinstance(control,
-                                      dace.graph.edges.LoopAssignment):
+                        if isinstance(control, cflow.LoopAssignment):
                             # Generate the transition, but leave the
                             # assignments to the loop
                             generate_transition &= True
                             generate_assignments &= False
 
-                        elif isinstance(control, dace.graph.edges.LoopBack):
+                        elif isinstance(control, cflow.LoopBack):
                             generate_transition &= False
                             generate_assignments &= False
 
-                        elif isinstance(control, dace.graph.edges.LoopExit):
+                        elif isinstance(control, cflow.LoopExit):
                             # Need to strip the condition, so generate it from
                             # the loop entry
                             generate_transition &= False
                             generate_assignments &= True
 
-                        elif isinstance(control, dace.graph.edges.LoopEntry):
+                        elif isinstance(control, cflow.LoopEntry):
                             generate_transition &= False
                             generate_assignments &= False
 
@@ -492,11 +492,11 @@ DACE_EXPORTED void __dace_exit_%s(%s)
                                         control.scope.exit.edge.dst))
                                 generated_edges.add(control.scope.exit.edge)
 
-                        elif isinstance(control, dace.graph.edges.IfExit):
+                        elif isinstance(control, cflow.IfExit):
                             generate_transition &= True
                             generate_assignments &= True
 
-                        elif isinstance(control, dace.graph.edges.IfEntry):
+                        elif isinstance(control, cflow.IfEntry):
                             generate_transition &= False
                             generate_assignments &= True
 
@@ -595,7 +595,7 @@ DACE_EXPORTED void __dace_exit_%s(%s)
                 # End of out_edges loop
 
             if (((len(out_edges) == 0) or
-                 (not isinstance(scope, dace.graph.edges.ControlFlowScope) and
+                 (not isinstance(scope, cflow.ControlFlowScope) and
                   (len(states_to_generate) == 0)))
                     and (len(states_generated) != sdfg.number_of_nodes())):
                 callsite_stream.write(
@@ -747,17 +747,15 @@ DACE_EXPORTED void __dace_exit_%s(%s)
                 # Nested loops case I - previous edge of internal loop is a
                 # loop-entry of an external loop (first state in a loop is
                 # another loop)
-                if (len(control_flow[previous_edge]) == 1
-                        and isinstance(control_flow[previous_edge][0],
-                                       dace.graph.edges.LoopEntry)):
+                if (len(control_flow[previous_edge]) == 1 and isinstance(
+                        control_flow[previous_edge][0], cflow.LoopEntry)):
                     # Nested loop, mark parent scope
                     loop_parent = control_flow[previous_edge][0].scope
                 # Nested loops case II - exit edge of internal loop is a
                 # back-edge of an external loop (last state in a loop is another
                 # loop)
-                elif (len(control_flow[exit_edge]) == 1
-                      and isinstance(control_flow[exit_edge][0],
-                                     dace.graph.edges.LoopBack)):
+                elif (len(control_flow[exit_edge]) == 1 and isinstance(
+                        control_flow[exit_edge][0], cflow.LoopBack)):
                     # Nested loop, mark parent scope
                     loop_parent = control_flow[exit_edge][0].scope
                 elif (len(control_flow[exit_edge]) == 0
@@ -778,7 +776,7 @@ DACE_EXPORTED void __dace_exit_%s(%s)
                     continue
 
                 # This is a loop! Generate the necessary annotation objects.
-                loop_scope = dace.graph.edges.LoopScope(internal_nodes)
+                loop_scope = cflow.LoopScope(internal_nodes)
 
                 if ((len(previous_edge.data.assignments) > 0
                      or len(back_edge.data.assignments) > 0) and
@@ -787,15 +785,14 @@ DACE_EXPORTED void __dace_exit_%s(%s)
                       control_flow[previous_edge][0].scope == loop_parent))):
                     # Generate assignment edge, if available
                     control_flow[previous_edge].append(
-                        dace.graph.edges.LoopAssignment(
-                            loop_scope, previous_edge))
+                        cflow.LoopAssignment(loop_scope, previous_edge))
                 # Assign remaining control flow constructs
                 control_flow[entry_edge].append(
-                    dace.graph.edges.LoopEntry(loop_scope, entry_edge))
+                    cflow.LoopEntry(loop_scope, entry_edge))
                 control_flow[exit_edge].append(
-                    dace.graph.edges.LoopExit(loop_scope, exit_edge))
+                    cflow.LoopExit(loop_scope, exit_edge))
                 control_flow[back_edge].append(
-                    dace.graph.edges.LoopBack(loop_scope, back_edge))
+                    cflow.LoopBack(loop_scope, back_edge))
 
             ###################################################################
             # If/then/else detection procedure
@@ -860,39 +857,34 @@ DACE_EXPORTED void __dace_exit_%s(%s)
                     continue
 
                 # This is a valid if/then/else construct. Generate annotations
-                if_then_else = dace.graph.edges.IfThenElse(
-                    candidate, dominator)
+                if_then_else = cflow.IfThenElse(candidate, dominator)
 
                 # Arbitrarily assign then/else to the two branches. If one edge
                 # has no dominator but leads to the dominator, it means there's
                 # only a then clause (and no else).
                 has_else = False
                 if len(dominators[left]) == 1:
-                    then_scope = dace.graph.edges.IfThenScope(
-                        if_then_else, left_nodes)
-                    else_scope = dace.graph.edges.IfElseScope(
-                        if_then_else, right_nodes)
+                    then_scope = cflow.IfThenScope(if_then_else, left_nodes)
+                    else_scope = cflow.IfElseScope(if_then_else, right_nodes)
                     control_flow[left_entry].append(
-                        dace.graph.edges.IfEntry(then_scope, left_entry))
+                        cflow.IfEntry(then_scope, left_entry))
                     control_flow[left_exit].append(
-                        dace.graph.edges.IfExit(then_scope, left_exit))
+                        cflow.IfExit(then_scope, left_exit))
                     control_flow[right_exit].append(
-                        dace.graph.edges.IfExit(else_scope, right_exit))
+                        cflow.IfExit(else_scope, right_exit))
                     if len(dominators[right]) == 1:
                         control_flow[right_entry].append(
-                            dace.graph.edges.IfEntry(else_scope, right_entry))
+                            cflow.IfEntry(else_scope, right_entry))
                         has_else = True
                 else:
-                    then_scope = dace.graph.edges.IfThenScope(
-                        if_then_else, right_nodes)
-                    else_scope = dace.graph.edges.IfElseScope(
-                        if_then_else, left_nodes)
+                    then_scope = cflow.IfThenScope(if_then_else, right_nodes)
+                    else_scope = cflow.IfElseScope(if_then_else, left_nodes)
                     control_flow[right_entry].append(
-                        dace.graph.edges.IfEntry(then_scope, right_entry))
+                        cflow.IfEntry(then_scope, right_entry))
                     control_flow[right_exit].append(
-                        dace.graph.edges.IfExit(then_scope, right_exit))
+                        cflow.IfExit(then_scope, right_exit))
                     control_flow[left_exit].append(
-                        dace.graph.edges.IfExit(else_scope, left_exit))
+                        cflow.IfExit(else_scope, left_exit))
 
         #######################################################################
         # Generate actual program body
